Solve challenge 1.3 and update XOR ciphers
Updates the XOR cipher implementation to include 2 more variants:
bytewise (with a single byte as a key), and repeating (a single
repeating key), and adds a function to compute the statistical
difference of a buffer using a probability distribution table by
calculating the character frequency in the buffer.

The solution for challenge 1.3 uses Bhattacharyya distance with a table
of English character probabilities (with an included space character, to
make sure the correct result is returned) to calculate a number in
the range of [0,1] where 1 = a perfect match.